restructure is_identified_vec_of moved to seperate file
